Fixed float arithmetic in get_time_ms.

+++ misc.c
@@ -1,7 +1,7 @@
+#include "defs.h"
 #include <stdio.h>
 #include <time.h>
-#include "defs.h"
 
 long get_time_ms(void) {
-	return (long) ((clock() / CLOCKS_PER_SEC ) * 1000);
+	return (long) (((float) clock() / CLOCKS_PER_SEC) * 1000.0f);
 }
